Move Tracer up a level; instana/tracer
