+++ lib/instana/tracer.rb
